Move methods from BaseCRUD to RequestMethods
Methods moved:
 * service_address
 * endpoint
 * entity_name
